GUILayers are now strings instead of unsigned ints. Replaced types where needed. Created static const strings to replace old #defined int values. Added GULayer to the labelMessage - labels should be attached to a layer.

+++ src/libwatcher/colorMessage.h
@@ -14,8 +14,8 @@ namespace watcher {
                 // The data
                 Color color;
                 boost::asio::ip::address nodeAddr;
-                uint32_t flashPeriod;        // this should be a float
-                uint32_t expiration;         // this should be a float
+                float flashPeriod; 
+                float expiration;  
 
                 ColorMessage();  // default: black, 127.0.0.1, no expire, no flash
                 ColorMessage(
